+++ apps/mobile/src/App.tsx
@@ -17,7 +17,6 @@ import { ThreadArrangementHost } from "./features/threads/ThreadArrangementSheet
 import { ConfirmDialogHost } from "./components/ConfirmDialogHost";
 import { InfinitusAlarmsBridge } from "./features/infinitus/InfinitusAlarmsBridge";
 import { InfinitusAlertPushBridge } from "./features/infinitus/InfinitusAlertPushBridge";
-import { InfinitusLiveActivityBridge } from "./features/infinitus/InfinitusLiveActivityBridge";
 import { InfinitusNotificationPresenter } from "./features/infinitus/InfinitusNotificationPresenter";
 import { CloudAuthProvider } from "./features/cloud/CloudAuthProvider";
 import { prepareNativeShowcaseCapture } from "./features/showcase/nativeShowcaseScene";
@@ -131,7 +130,6 @@ function AppContent() {
               </IncomingShareProvider>
               <ConfirmDialogHost />
               <ThreadArrangementHost />
-              <InfinitusLiveActivityBridge />
               <InfinitusAlarmsBridge />
               <InfinitusAlertPushBridge />
               <InfinitusNotificationPresenter />